Add SPDX License identifiers and be honest in MODULE_LICENSE
According to https://docs.kernel.org/process/license-rules.html#id1
"GPL" as parameter to MODULE_LICENSE() means "Module is licensed under
GPL version 2". This is not true for this source code which is GPL v3 or
later.

So mark the code with the SPDX license identifier GPL-3.0+ and fix the
MODULE_LICENSE() macro usages accordingly.

Note that this change enforces license compilance (at least a bit)
because some kernel code is only available to modules with a GPL-2.0
compatible license.